requirements.txt
opencv-python
mediapipe
boto3
python-dotenv
numpy
fastapi
uvicorn
jinja2
